northd: Fix logical router load-balancer nat rules when using DGP.
This commit fixes the build_distr_lrouter_nat_flows_for_lb function to
include a DNAT flow entry for each DGP in use. Since we have added support
to create multiple gateway ports per logical router, it's necessary to
include in the LR NAT rules pipeline a specific entry for each attached DGP.
Otherwise, the inbound traffic will only be redirected when the incoming LRP
matches the chassis_resident field.

Additionally, this patch includes the ability to use load-balancer with DGPs
attached to multiple chassis. We can have each of the DGPs associated with a
different chassis, and in this case the DNAT rules added by default will not
be enough to guarantee outgoing traffic.

To solve the multiple chassis for DGPs problem, this patch include a new
config options to be configured in the load-balancer. If the use_stateless_nat
is set to true, the logical router that references this load-balancer will use
Stateless NAT rules when the logical router has multiple DGPs. After applying
this patch and setting the use_stateless_nat option, the inbound and/or
outbound traffic can pass through any chassis where the DGP resides without
having problems with CT state.

Reported-at: https://bugs.launchpad.net/ubuntu/+source/ovn/+bug/2054322
Fixes: 15348b7b806f ("ovn-northd: Multiple distributed gateway port support.")

Signed-off-by: Roberto Bartzen Acosta <rbartzen@gmail.com>